backend/requirements.txt
# =============================================================================
# Zonara Backend — Python Dependencies
# @date    25 Maret 2026
# @version 1.0
# =============================================================================

# --- Framework & Server ---
fastapi==0.110.0
uvicorn[standard]==0.27.1

# --- Database (Async) ---
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# --- Validation & Settings ---
pydantic==2.6.1
pydantic-settings==2.1.0

# --- Authentication (JWT + Password Hashing) ---
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# --- File Upload & Form Parsing ---
python-multipart==0.0.9

# --- QR Code Generation ---
qrcode[pil]==7.4.2
Pillow==10.2.0

# --- PDF Report Generation ---
reportlab==4.1.0

# --- WebSocket ---
websockets==12.0

# --- Testing ---
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0